Moved packages, added launch activity
