+++ sample/src/main/kotlin/me/dmdev/rxpm/sample/App.kt
@@ -1,15 +1,13 @@
 package me.dmdev.rxpm.sample
 
-import android.annotation.SuppressLint
 import android.app.Application
-
+import me.dmdev.rxpm.sample.main.AppComponent
 import timber.log.Timber
 
 
 class App : Application() {
 
     companion object {
-        @SuppressLint("StaticFieldLeak")
         lateinit var component: AppComponent
             private set
     }
