--- sample/src/main/kotlin/me/dmdev/rxpm/sample/extensions/AppExtensions.kt
@@ -1,3 +1,0 @@
-package me.dmdev.rxpm.sample.extensions
-
-fun String.onlyDigits() = this.replace("\\D".toRegex(), "")
