+++ sample/src/main/kotlin/me/dmdev/rxpm/sample/main/extensions/AppExtensions.kt
@@ -1,0 +1,3 @@
+package me.dmdev.rxpm.sample.main.extensions
+
+fun String.onlyDigits() = this.replace("\\D".toRegex(), "")
