+++ sample/src/main/kotlin/me/dmdev/rxpm/sample/main/model/TokenStorage.kt
@@ -1,0 +1,21 @@
+package me.dmdev.rxpm.sample.main.model
+
+import java.util.concurrent.atomic.AtomicReference
+
+
+class TokenStorage {
+
+    private var tokenRef = AtomicReference<String>("")
+
+    fun saveToken(token: String) {
+        tokenRef.set(token)
+    }
+
+    fun getToken(): String {
+        return tokenRef.get()
+    }
+
+    fun clear() {
+        tokenRef.set("")
+    }
+}
