+++ sample/src/main/kotlin/me/dmdev/rxpm/sample/main/ui/base/BackHandler.kt
@@ -1,0 +1,6 @@
+package me.dmdev.rxpm.sample.main.ui.base
+
+
+interface BackHandler {
+    fun handleBack(): Boolean
+}
