+++ sample/src/main/kotlin/me/dmdev/rxpm/sample/main/util/ResourcesProvider.kt
@@ -1,0 +1,12 @@
+package me.dmdev.rxpm.sample.main.util
+
+import android.content.Context
+import android.support.annotation.StringRes
+
+
+class ResourceProvider(private val context: Context) {
+
+    fun getString(@StringRes resId: Int, vararg formatArgs: Any): String {
+        return context.resources.getString(resId, *formatArgs)
+    }
+}
