--- sample/src/main/kotlin/me/dmdev/rxpm/sample/ui/base/BackHandler.kt
@@ -1,6 +1,0 @@
-package me.dmdev.rxpm.sample.ui.base
-
-
-interface BackHandler {
-    fun handleBack(): Boolean
-}
