--- sample/src/main/kotlin/me/dmdev/rxpm/sample/ui/base/ScreenPresentationModel.kt
@@ -1,32 +1,0 @@
-package me.dmdev.rxpm.sample.ui.base
-
-import me.dmdev.rxpm.PresentationModel
-import me.dmdev.rxpm.navigation.NavigationMessage
-import me.dmdev.rxpm.sample.BackMessage
-import me.dmdev.rxpm.widget.dialogControl
-
-
-abstract class ScreenPresentationModel : PresentationModel() {
-
-    val errorDialog = dialogControl<String, Unit>()
-
-    private val backActionDefault = Action<Unit>()
-
-    open val backAction: Action<Unit>  = backActionDefault
-
-    override fun onCreate() {
-        super.onCreate()
-
-        backActionDefault.observable
-                .subscribe { sendMessage(BackMessage()) }
-                .untilDestroy()
-    }
-
-    protected fun sendMessage(message: NavigationMessage) {
-        navigationMessages.consumer.accept(message)
-    }
-
-    protected fun showError(errorMessage: String?) {
-        errorDialog.show(errorMessage ?: "Unknown error")
-    }
-}
