+++ sample/src/main/res/layout/activity_main.xml
@@ -5,4 +5,4 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="me.dmdev.rxpm.sample.AppActivity"/>
+    tools:context="me.dmdev.rxpm.sample.main.AppActivity"/>
